requirements.txt
absl-py==2.1.0
accelerate==0.32.1
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
Brotli 
cachetools==5.3.3
certifi==2023.5.7
charset-normalizer 
click==8.1.4
clip 
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
decorator==5.1.1
docker-pycreds==0.4.0
eval_type_backport==0.2.0
executing==1.2.0
filelock==3.15.4
fonttools==4.43.1
fsspec==2024.6.1
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.31.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
h5py==3.10.0
huggingface-hub==0.23.4
idna
imageio==2.34.1
importlib-resources==6.1.0
importlib_metadata==8.0.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
joblib==1.3.1
kiwisolver==1.4.5
kornia==0.6.12
large-image==1.29.2
large-image-source-openslide==1.29.2
lazy_loader==0.4
lightning-utilities==0.10.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.6
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.1
nibabel==5.2.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
openslide-python==1.3.1
openTSNE==1.0.1
packaging==23.1
palettable==3.3.3
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
peft==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
pyparsing==3.1.1
PySocks 
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.27.1
setproctitle==1.3.2
SimpleITK==2.3.0
six
smmap==5.0.0
stack-data==0.6.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
tifftools==1.5.2
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchmetrics==1.2.1
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3
wandb==0.15.5
wcwidth==0.2.6
Werkzeug==3.0.3
zipp==3.17.0
